Common Pages Requires Different Layout. Boxes.scss for flexbox
Common Page - which are used in the application before logging in.
Have different usage of the menu at the top. ERB is ugly
and confusing for the account_signed_in? and I'm trying to see
if I can avoid that by separating unsigned pages from signed
in pages.

Boxes - is a basic flexbox usage. We'll see how it develops.
Objects.* is used for css that does not have visible UI to it.