Bugfix removed extended wrapped elements to full browser width
Extended wrapped elements was a hack to expand the header outside
the width of the body container. This required a number of css
rules including large margins and paddings and making overflow
hidden. All came undone as mobile wants to display the extended
margins.

Should have just done was add a container div for content. More
wrappers but simpler code.