OfferSearchTest Replaces the generic BookSearchTest. Remove Book Index
Originally BookSearchTest was to cover the usage of the search
button in Offer, Books and Want. However, they are slightly
different and anyway you shouldn't feel the necessessity to be
dry in tests in the same manner as production code.

Book isn't going to be used (it will become trade) I'm going to
blank it out to avoid maintenance of a page I don't intend to use