build_offer_trade, build_want_trade were matching completed trades
Bug
Account 1 creates a want for a book.
Account 2 creats an offer for a book.
Application creates a trade.
Account 1 creates a want for the same book.

Expect: Nothing to happen
Action: Application creats a trade - matching the previous offer
        on the book

This fix solves it in the application but I couldn't reproduce it
in the model tests.